requirements.txt
click==8.3.0
colorama==0.4.6
moviepy==2.2.1
yt_dlp==2025.10.22
pydantic>=2.0
whisper>=1.0
ffmpeg-python>=0.2.0
boto3>=1.34.0
openai>=1.0
anthropic>=0.77.0
tiktoken>=0.5.0
pyyaml>=6.0
jinja2>=3.0
requests>=2.31.0
